+++ git-cliff-core/src/contributor.rs
